delete contact operations done

+++ Controllers/PeopleController.cs
@@ -16,7 +16,7 @@ namespace phoneBook.Controllers
         [HttpGet]
         public ActionResult ListPeople(int id)
         {
-            return View(db.peopleInDirectory.Where(x => x.userId == id).ToList());
+                return View(db.peopleInDirectory.Where(x => x.userId == id).ToList());
         }
 
         public ActionResult Create()
@@ -90,13 +90,18 @@ namespace phoneBook.Controllers
                 return View();
             }
         }
-        [HttpPost]
+
+        [HttpGet]
         public ActionResult Delete(int id)
         {
             var p = db.peopleInDirectory.FirstOrDefault(x => x.id == id);
-            db.peopleInDirectory.Remove(p);
-            db.SaveChanges();
-            return RedirectToAction("ListPeople", "People", new { id = p.userId });
+            int userId = p.userId;
+            if(p !=null)
+            {
+                db.peopleInDirectory.Remove(p);
+                db.SaveChanges();
+            }
+            return RedirectToAction("ListPeople", "People", new { id = userId });
         }
 
     }
